dist: add COPYRIGHT*.html files to distribution
Also adds contents of LICENSE folder because some of our in-tree licences will require that the license text is reproduced. The wording of COPYRIGHT is adjusted appropriately, and the top-level LICENSE-APACHE and LICENSE-MIT files are removed.

There's a LICENSE.rtf generated as part of the Windows MSI installer - that now takes the text from the LICENSES folder.